refactor: move RateType into common module
libecalc.common should not import from libecalc.dto
