+++ src/libecalc/core/consumers/legacy_consumer/consumer_function/direct_expression_consumer_function.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 import numpy as np
 from libecalc import dto
 from libecalc.common.units import Unit
-from libecalc.common.utils.rates import Rates
+from libecalc.common.utils.rates import Rates, RateType
 from libecalc.core.consumers.legacy_consumer.consumer_function import (
     ConsumerFunction,
     ConsumerFunctionResult,
@@ -15,7 +15,6 @@ from libecalc.core.consumers.legacy_consumer.consumer_function.utils import (
     get_power_loss_factor_from_expression,
 )
 from libecalc.core.models.results import EnergyFunctionGenericResult
-from libecalc.dto.types import RateType
 from libecalc.dto.variables import VariablesMap
