+++ src/tests/libecalc/core/consumers/conftest.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 
+import libecalc.common.utils.rates
 import pytest
 from libecalc import dto
 from libecalc.dto import VariablesMap
@@ -83,22 +84,22 @@ def direct_el_consumer() -> dto.ElectricityConsumer:
             datetime(2020, 1, 1): dto.DirectConsumerFunction(
                 load=Expression.setup_from_expression(value=1),
                 energy_usage_type=dto.types.EnergyUsageType.POWER,
-                consumption_rate_type=dto.types.RateType.STREAM_DAY,
+                consumption_rate_type=libecalc.common.utils.rates.RateType.STREAM_DAY,
             ),
             datetime(2021, 1, 1): dto.DirectConsumerFunction(  # Run above capacity
                 load=Expression.setup_from_expression(value=2),
                 energy_usage_type=dto.types.EnergyUsageType.POWER,
-                consumption_rate_type=dto.types.RateType.STREAM_DAY,
+                consumption_rate_type=libecalc.common.utils.rates.RateType.STREAM_DAY,
             ),
             datetime(2022, 1, 1): dto.DirectConsumerFunction(  # Run above capacity
                 load=Expression.setup_from_expression(value=10),
                 energy_usage_type=dto.types.EnergyUsageType.POWER,
-                consumption_rate_type=dto.types.RateType.STREAM_DAY,
+                consumption_rate_type=libecalc.common.utils.rates.RateType.STREAM_DAY,
             ),
             datetime(2023, 1, 1): dto.DirectConsumerFunction(  # Ensure we handle 0 load as well.
                 load=Expression.setup_from_expression(value=0),
                 energy_usage_type=dto.types.EnergyUsageType.POWER,
-                consumption_rate_type=dto.types.RateType.STREAM_DAY,
+                consumption_rate_type=libecalc.common.utils.rates.RateType.STREAM_DAY,
             ),
         },
         regularity={datetime(1900, 1, 1): Expression.setup_from_expression(1)},
